Finished setup and tests for error cases on item scraping
Ok, this feature is functioning. Putting in a
valid Ikea product URL scrapes info from the site
and creates an AR entry for it. 

We now avoid most bad URL use cases, and handle 
duplicate URL attempt by forwarding the user to
the already-created entry for that product.

Changes:
- Added URL validation - some edge cases likely 
not covered
- Now validating product URLs are unique.
- Set up Factory Girl
- Some tests added for scraping items, invalid 
URLs, duplicate items
- Unit test for Item get accessor - needed one to
return two digits on the item price